설정 수정 + 충돌횟수

+++ Assets/Scripts/GameManager.cs
@@ -2,14 +2,15 @@
 using System.Collections.Generic;
 using UnityEngine;
 using UnityEngine.SceneManagement;
-
+using UnityEngine.SocialPlatforms.Impl;
+using UnityEngine.UI;
 
 public class GameManager : MonoSingleton<GameManager>
 {
     public bool downCheck = false;
 
     public float speed = 1000f;
-    public float CCount = 0f;
+    public int CCount = 0;
     public int currentLevel = 2;
 
     public StageData saveJson;
@@ -27,6 +28,7 @@ public class GameManager : MonoSingleton<GameManager>
             speed = 1000f;
             SceneManager.LoadScene(SceneManager.GetActiveScene().buildIndex);
         }
+
     }
 
 }
